fix(api): keep runtime observation v1 frozen
