+++ backend/app/schemas/runtime_observed.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 from datetime import UTC, datetime
 from typing import Literal
-from uuid import UUID
 
 from pydantic import (
     BaseModel,
@@ -15,13 +14,6 @@ from pydantic import (
 )
 
 RuntimeObservedStatus = Literal["ok", "error", "unknown"]
-AgentPluginObservedStatus = Literal["installed", "failed", "unknown"]
-AgentPluginObservationErrorCode = Literal[
-    "reconcile_failed",
-    "receipt_missing",
-    "receipt_unreadable",
-    "receipt_mismatch",
-]
 
 
 class _StrictObservedWireModel(BaseModel):
@@ -112,102 +104,6 @@ class HostedRuntimeObservedAppliedV2(_StrictObservedWireModel):
         return value
 
 
-class HostedRuntimeObservedAgentPluginV1(_StrictObservedWireModel):
-    installation_id: str = Field(alias="installationId", min_length=1, max_length=200)
-    name: str = Field(
-        min_length=1,
-        max_length=64,
-        pattern=r"^[a-z0-9][a-z0-9.-]{0,63}$",
-    )
-    version: str = Field(
-        min_length=1,
-        max_length=256,
-        pattern=(
-            r"^(?:0|[1-9][0-9]*)\.(?:0|[1-9][0-9]*)\.(?:0|[1-9][0-9]*)"
-            r"(?:-(?:0|[1-9][0-9]*|[0-9A-Za-z-]*[A-Za-z-][0-9A-Za-z-]*)"
-            r"(?:\.(?:0|[1-9][0-9]*|[0-9A-Za-z-]*[A-Za-z-][0-9A-Za-z-]*))*)?"
-            r"(?:\+[0-9A-Za-z-]+(?:\.[0-9A-Za-z-]+)*)?$"
-        ),
-    )
-    content_digest: str = Field(
-        alias="contentDigest",
-        pattern=r"^sha256-tree-v1:[0-9a-f]{64}$",
-    )
-    source_revision: str = Field(alias="sourceRevision", pattern=r"^[0-9a-f]{64}$")
-    generation: int = Field(ge=1, le=9_007_199_254_740_991)
-    status: AgentPluginObservedStatus
-    error_code: AgentPluginObservationErrorCode | None = Field(
-        alias="errorCode",
-        default=None,
-    )
-
-    @field_validator("installation_id")
-    @classmethod
-    def validate_installation_id(cls, value: str) -> str:
-        try:
-            parsed = UUID(value)
-        except ValueError as exc:
-            raise ValueError("installationId must be a canonical UUID") from exc
-        if str(parsed) != value:
-            raise ValueError("installationId must be a canonical UUID")
-        return value
-
-    @field_validator("name")
-    @classmethod
-    def validate_name(cls, value: str) -> str:
-        if "--" in value or ".." in value or value[-1] in {".", "-"}:
-            raise ValueError("name must be a canonical Agent Plugin name")
-        return value
-
-    @model_validator(mode="after")
-    def validate_status_error(self) -> HostedRuntimeObservedAgentPluginV1:
-        if self.status == "installed" and self.error_code is not None:
-            raise ValueError("installed Agent Plugin observation cannot include errorCode")
-        if self.status == "failed" and self.error_code != "reconcile_failed":
-            raise ValueError("failed Agent Plugin observation requires reconcile_failed")
-        if self.status == "unknown" and self.error_code not in {
-            "receipt_missing",
-            "receipt_unreadable",
-            "receipt_mismatch",
-        }:
-            raise ValueError("unknown Agent Plugin observation requires a receipt error code")
-        return self
-
-
-class HostedRuntimeObservedAgentPluginsV1(_StrictObservedWireModel):
-    schema_version: Literal[1] = Field(alias="schemaVersion")
-    installations: list[HostedRuntimeObservedAgentPluginV1] = Field(max_length=128)
-
-    @field_validator("installations")
-    @classmethod
-    def validate_installations(
-        cls,
-        value: list[HostedRuntimeObservedAgentPluginV1],
-    ) -> list[HostedRuntimeObservedAgentPluginV1]:
-        names = [installation.name for installation in value]
-        installation_ids = [installation.installation_id for installation in value]
-        if len(names) != len(set(names)) or len(installation_ids) != len(set(installation_ids)):
-            raise ValueError("Agent Plugin observations must have unique identities")
-        if names != sorted(names, key=lambda name: name.encode("utf-8")):
-            raise ValueError("Agent Plugin observations must be sorted by name")
-        return value
-
-    def validate_applied_identity(self, applied: HostedRuntimeObservedAppliedV2) -> None:
-        for installation in self.installations:
-            if installation.status == "failed":
-                if installation.generation < applied.generation or (
-                    installation.generation == applied.generation
-                    and installation.source_revision != applied.source_revision
-                ):
-                    raise ValueError("failed Agent Plugin observation is stale")
-                continue
-            if (
-                installation.generation != applied.generation
-                or installation.source_revision != applied.source_revision
-            ):
-                raise ValueError("Agent Plugin observation must match applied identity")
-
-
 class HostedRuntimeObservedV2(_StrictObservedWireModel):
     schema_version: Literal["clawdi.hostedRuntimeObserved.v2"] = Field(alias="schemaVersion")
     reported_at: datetime = Field(alias="reportedAt")
@@ -224,10 +120,6 @@ class HostedRuntimeObservedV2(_StrictObservedWireModel):
     systemd: HostedRuntimeObservedSystemdV1 | None = None
     supervisor: HostedRuntimeObservedSupervisorV1 | None = None
     providers: dict[str, HostedRuntimeObservedProviderPayload] | None = None
-    agent_plugins: HostedRuntimeObservedAgentPluginsV1 | None = Field(
-        alias="agentPlugins",
-        default=None,
-    )
     error: str | None = Field(default=None, max_length=4000)
     converge_error: str | None = Field(alias="convergeError", default=None, max_length=4000)
     truncated: bool | None = None
@@ -247,15 +139,6 @@ class HostedRuntimeObservedV2(_StrictObservedWireModel):
         if parsed.tzinfo is None:
             raise ValueError("reportedAt must include a timezone")
         return parsed.astimezone(UTC)
-
-    @model_validator(mode="after")
-    def validate_agent_plugin_identity(self) -> HostedRuntimeObservedV2:
-        if self.agent_plugins is None:
-            return self
-        if self.applied is None:
-            raise ValueError("Agent Plugin observation requires applied identity")
-        self.agent_plugins.validate_applied_identity(self.applied)
-        return self
 
 
 HostedRuntimeObserved = HostedRuntimeObservedV2
